Refuse category removal if there are events linked to this category

+++ main-service/src/main/java/ru/practicum/service/CategoryServiceImpl.java
@@ -7,6 +7,7 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
+import ru.practicum.exceptions.ConflictException;
 import ru.practicum.storage.CategoryRepository;
 import ru.practicum.dto.category.CategoryDto;
 import ru.practicum.dto.category.NewCategoryDto;
@@ -21,6 +22,8 @@ import java.util.List;
 @RequiredArgsConstructor
 public class CategoryServiceImpl implements CategoryService {
     private final CategoryRepository categoryRepository;
+
+    private final EventService eventService;
 
     @Override
     public List<CategoryDto> findPortion(Integer from, Integer size) {
@@ -49,10 +52,12 @@ public class CategoryServiceImpl implements CategoryService {
     @Transactional
     public CategoryDto delete(Long id) {
         final Category category = getCategoryIfExists(id);
+        if (eventService.countEventByCategory(id) > 0) {
+            throw new ConflictException("Не найдена категория с ID %d".formatted(id));
+        }
         categoryRepository.deleteById(id);
         log.info("Удалена категория с ID {}, ее содержимое {}", id, category);
         return CategoryMapper.INSTANCE.toCategoryDto(category);
-
     }
 
     @Override
